% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcreporter.R
\name{funcreporter}
\alias{funcreporter}
\title{Generate a parameterized report from template}
\usage{
funcreporter(
  template_name,
  output_format = "html_document",
  output_file,
  params = NULL,
  remove_copied_template_files = TRUE,
  envir = new.env(),
  ...
)
}
\arguments{
\item{template_name}{Template name from package.}

\item{output_format}{The R Markdown output format to convert to, e.g. "html_document", "pdf_document".}

\item{output_file}{The name of the output file. Non-existent directories will be created recursively.}

\item{params}{A list of named parameters.}

\item{remove_copied_template_files}{An option to remove copied template files after rendering.}

\item{envir}{The environment in which the code chunks are to be evaluated during knitting.}

\item{...}{Other options forwarded to rmarkdown::render}
}
\description{
Generate a parameterized report from template
}
\examples{
\dontrun{
library(funcreporter)
set_funcreporter_pkg("funcreporter")
funcreporter(
    template_name = "Sample Template",
    output_format = "html_document",
    output_file = "versicolor-report",
    params = list(species = "versicolor"))
}
}
