% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_report_descriptions}
\alias{print_report_descriptions}
\title{Print YAML report descriptions to the console}
\usage{
print_report_descriptions()
}
\description{
Print YAML report descriptions to the console
}
\examples{
\dontrun{
print_report_descriptions()
}
}
