% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_funcreporter_pkg}
\alias{set_funcreporter_pkg}
\title{Set funcreporter env vars}
\usage{
set_funcreporter_pkg(pkg_name)
}
\arguments{
\item{pkg_name}{Name of package with templates}
}
\description{
Set funcreporter env vars
}
\examples{
\dontrun{
set_funcreporter_pkg("reporting_pacakge")
}
}
